+++ packages/widgetbook/lib/src/widgetbook.dart
@@ -25,6 +25,9 @@ class Widgetbook extends StatefulWidget {
     this.appBuilder = widgetsAppBuilder,
     this.addons,
     this.integrations,
+    this.lightTheme,
+    this.darkTheme,
+    this.themeMode,
   });
 
   /// A [Widgetbook] with [CupertinoApp] as an [appBuilder].
@@ -35,6 +38,9 @@ class Widgetbook extends StatefulWidget {
     this.appBuilder = cupertinoAppBuilder,
     this.addons,
     this.integrations,
+    this.lightTheme,
+    this.darkTheme,
+    this.themeMode,
   });
 
   /// A [Widgetbook] with [MaterialApp] as an [appBuilder].
@@ -45,6 +51,9 @@ class Widgetbook extends StatefulWidget {
     this.appBuilder = materialAppBuilder,
     this.addons,
     this.integrations,
+    this.lightTheme,
+    this.darkTheme,
+    this.themeMode,
   });
 
   /// The initial route for that will be used on first startup.
@@ -70,6 +79,22 @@ class Widgetbook extends StatefulWidget {
   /// integrate with Widgetbook Cloud via [WidgetbookCloudIntegration], but
   /// can also be used to integrate with third-party packages.
   final List<WidgetbookIntegration>? integrations;
+
+  /// The custom theme for the Widgetbook interface when using light mode.
+  ///
+  /// This theme will override the default light theme provided by Widgetbook.
+  final ThemeData? lightTheme;
+
+  /// The custom theme for the Widgetbook interface when using dark mode.
+  ///
+  /// This theme will override the default dark theme provided by Widgetbook.
+  final ThemeData? darkTheme;
+
+  /// The theme mode to be applied to the Widgetbook application.
+  ///
+  /// This parameter allows you to set the theme to light, dark, or follow the system setting.
+  /// By default, it follows the system theme.
+  final ThemeMode? themeMode;
 
   @override
   State<Widgetbook> createState() => _WidgetbookState();
@@ -110,8 +135,9 @@ class _WidgetbookState extends State<Widgetbook> {
       state: state,
       child: MaterialApp.router(
         title: 'Widgetbook',
-        theme: Themes.light,
-        darkTheme: Themes.dark,
+        themeMode: widget.themeMode ?? ThemeMode.system,
+        theme: widget.lightTheme ?? Themes.light,
+        darkTheme: widget.darkTheme ?? Themes.dark,
         routerConfig: router,
         debugShowCheckedModeBanner: false,
       ),